+++ lib/mongoid/validations.rb
@@ -44,11 +44,14 @@ module Mongoid #:nodoc:
     #
     # @since 2.0.0.rc.1
     def read_attribute_for_validation(attr)
-      if relations[attr.to_s]
+      attribute = attr.to_s
+      if relations[attribute]
         begin_validate
         relation = send(attr)
         exit_validate
         relation.do_or_do_not(:in_memory) || relation
+      elsif fields[attribute] && fields[attribute].localized?
+        attributes[attribute]
       else
         send(attr)
       end
@@ -119,7 +122,20 @@ module Mongoid #:nodoc:
       def validates_uniqueness_of(*args)
         validates_with(UniquenessValidator, _merge_attributes(args))
       end
-      
+
+      # Validates whether or not a field is present - meaning nil or empty.
+      #
+      # @example
+      #   class Person
+      #     include Mongoid::Document
+      #     field :title
+      #
+      #     validates_presence_of :title
+      #   end
+      #
+      # @param [ Array ] args The names of the fields to validate.
+      #
+      # @since 2.4.0
       def validates_presence_of(*args)
         validates_with(PresenceValidator, _merge_attributes(args))
       end