+++ spec/functional/mongoid/validations/presence_spec.rb
@@ -2,15 +2,117 @@ require "spec_helper"
 
 describe Mongoid::Validations do
 
-  let(:updated_products) do
-    [ "Laptop", "Tablet", "Smartphone", "Desktop" ]
+  before do
+    [ Manufacturer, Product ].each(&:delete_all)
   end
 
-  let(:manufacturer) do
-    Manufacturer.create!(:products => [ "Laptop", "Tablet" ])
+  context "when validating a localized field" do
+
+    context "when any translation is blank" do
+
+      let(:product) do
+        Product.new
+      end
+
+      before do
+        product.name_translations = { "de" => "" }
+      end
+
+      it "is not a valid document" do
+        product.should_not be_valid
+      end
+
+      it "includes the proper errors" do
+        product.valid?
+        product.errors[:name].should_not be_empty
+      end
+    end
+
+    context "when any translation is nil" do
+
+      let(:product) do
+        Product.new
+      end
+
+      before do
+        product.name_translations = { "de" => nil }
+      end
+
+      it "is not a valid document" do
+        product.should_not be_valid
+      end
+
+      it "includes the proper errors" do
+        product.valid?
+        product.errors[:name].should_not be_empty
+      end
+    end
+
+    context "when the entire field is nil" do
+
+      let(:product) do
+        Product.new
+      end
+
+      before do
+        product.name_translations = nil
+      end
+
+      it "is not a valid document" do
+        product.should_not be_valid
+      end
+
+      it "includes the proper errors" do
+        product.valid?
+        product.errors[:name].should_not be_empty
+      end
+    end
+
+    context "when the entire field is empty" do
+
+      let(:product) do
+        Product.new
+      end
+
+      before do
+        product.name_translations = {}
+      end
+
+      it "is not a valid document" do
+        product.should_not be_valid
+      end
+
+      it "includes the proper errors" do
+        product.valid?
+        product.errors[:name].should_not be_empty
+      end
+    end
+
+    context "when the translations are present" do
+
+      let(:product) do
+        Product.new
+      end
+
+      before do
+        product.name_translations = { "en" => "test" }
+      end
+
+      it "is a valid document" do
+        product.should be_valid
+      end
+    end
   end
 
   context "when presence_of array attribute is updated and saved" do
+
+    let(:updated_products) do
+      [ "Laptop", "Tablet", "Smartphone", "Desktop" ]
+    end
+
+    let(:manufacturer) do
+      Manufacturer.create!(:products => [ "Laptop", "Tablet" ])
+    end
 
     before do
       manufacturer.products = updated_products
@@ -27,6 +129,14 @@ describe Mongoid::Validations do
   end
 
   context "when an array attribute has been updated" do
+
+    let(:updated_products) do
+      [ "Laptop", "Tablet", "Smartphone", "Desktop" ]
+    end
+
+    let(:manufacturer) do
+      Manufacturer.create!(:products => [ "Laptop", "Tablet" ])
+    end
 
     context "when retrieved, flattened and iterated" do